requirements.txt
google-generativeai==0.5.2
grpcio-status==1.62.2
pandas==2.2.2
pillow==10.3.0
pymongo==4.6.3
reflex==0.4.8.post1
